style: Rename `src/*.ts` with kebab-case Recommended by eslint-unicorn
