+++ src/variable-resolver.ts
