+++ webpack.config.ts
@@ -48,7 +48,7 @@ const browserConfig: Configuration = {
     target: 'webworker',
     mode: 'none',
     entry: {
-        extensionWeb: './src/web/extensionWeb.ts',
+        extensionWeb: './src/web/extension-web.ts',
         // 'test/suite/index': './src/web/test/suite/index.ts',
     },
     output: {
